Use log destination param for pgsql
This commit ensures that `stderr` is used for logging, by
specifying the command line parameter.

On FreeBSD this is very important otherwise syslog will be used and
the db-fixtures will hang as it looks in the expected log file and
loops forever waiting for a "database is ready" entry to appear...
`log_destination=stderr` is default on most systems and can be set in
`postgresql.conf` or given as an command line argument.

Read more at
http://www.postgresql.org/docs/9.1/static/runtime-config-logging.html