PGsql connection error on FreeBSD jailed environments
In FreeBSD jailed environments the loopback interface can not be used to
connect to pgsql, because it points to the loopback interface of the host and
not the jail.

Thus, it seems to pgsql like the connection is comming from the ip address
assigned to the jail.

This commit ensures that pgsql can be connected to from any host, when
run on a FreeBSD systems. This this package can be used in FreeBSD
jailed environments

This commit/PR originated from #105